Add config for density of confetti; ensure that it's independent of other choices
Previously, the confetti density wasn’t configurable, and would vary
depending on the number of colors and shapes included. (For example, if
you have 1 color vs 8 colors, the visual density of the CheerView was
dramatically different).

Now, you can adjust the density of the confetti, and it’s independent
of whatever color and shape parameters you specify.